pyproject.toml
[tool.poetry]
name = "iot"
version = "0.1.0"
description = ""
authors = ["huytran"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.72.0"
asyncio = "^3.4.3"
influxdb-client = "^1.25.0"
paho-mqtt = "^1.6.1"
Jinja2 = "^3.0.3"
SQLAlchemy = "^1.4.31"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
uvicorn = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
